requirements.txt
Django==5.0.6
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
psycopg2-binary==2.9.9   # if using PostgreSQL
